Refactor: '롤링페이퍼에 메세지 보내기 페이지' 디렉토리 수정에 의한 경로 변경
